requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aioresponses==0.7.6
aiosignal==1.3.1
attrs==24.2.0
coverage==7.6.1
frozenlist==1.4.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
six==1.16.0
yarl==1.13.1
